Inline ModelTransformer inside DefaultModelBuilder
