+++ api/maven-api-spi/src/main/java/org/apache/maven/api/spi/ModelTransformer.java
@@ -23,7 +23,6 @@ import org.apache.maven.api.annotations.Experimental;
 import org.apache.maven.api.annotations.Nonnull;
 import org.apache.maven.api.di.Named;
 import org.apache.maven.api.model.Model;
-import org.apache.maven.api.services.ModelTransformerException;
 
 /**
  * Marker interface for model transformers.
