+++ maven-model-builder/src/main/java/org/apache/maven/model/building/DefaultTransformerContextBuilder.java
@@ -201,7 +201,7 @@ class DefaultTransformerContextBuilder implements TransformerContextBuilder {
     }
 
     public FileModelSource getSource(String groupId, String artifactId) {
-        Set<FileModelSource> sources = mappedSources.get(groupId + ":" + artifactId);
+        Set<FileModelSource> sources = mappedSources.get(groupId != null ? groupId + ":" + artifactId : artifactId);
         if (sources == null) {
             return null;
         }
@@ -218,5 +218,8 @@ class DefaultTransformerContextBuilder implements TransformerContextBuilder {
         mappedSources
                 .computeIfAbsent(groupId + ":" + artifactId, k -> new HashSet<>())
                 .add(source);
+        if (groupId != null) {
+            mappedSources.computeIfAbsent(artifactId, k -> new HashSet<>()).add(source);
+        }
     }
 }
